Refactor: Update npm commands. Move server code into src/.
