+++ README.md
@@ -6,5 +6,6 @@ First time running the project.
 `npm install`
 
 Run Development Servers in 2 seperate terminal tabs.
-`npm run develop`
-`npm run devWebpack`
+`server:watch`
+
+Go to http://127.0.0.1:8080/.
