+++ package.json
@@ -2,12 +2,12 @@
   "name": "maptileserver",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "src/server/server.js",
   "scripts": {
-    "compile": "webpack",
-    "start": "node server.js",
-    "develop": "nodemon ./server.js",
-    "devWebpack": "webpack --d --watch",
+    "build": "webpack",
+    "server:watch": "nodemon --watch 'src/**/*' --exec 'npm run build && npm run start'",
+    "server:webpack": "webpack --d --watch",
+    "start": "node src/server/server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
